refactor: DailyChallengeScreen.kt decomposition
