+++ app/src/main/java/com/pv239/beelocal/ui/screens/dailychallenge/components/TemperatureLegend.kt
@@ -1,0 +1,68 @@
+package com.pv239.beelocal.ui.screens.dailychallenge.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun TemperatureLegend(currentTemperature: ProximityTemperature?) {
+    Card(
+        shape = RoundedCornerShape(16.dp),
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.surfaceContainer,
+        ),
+    ) {
+        Column {
+            ProximityTemperature.entries.forEachIndexed { index, temp ->
+                val isActive = temp == currentTemperature
+
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .background(if (isActive) temp.color.copy(alpha = 0.18f) else Color.Transparent)
+                        .padding(horizontal = 16.dp, vertical = 10.dp),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(12.dp),
+                ) {
+                    Text(text = temp.emoji, style = MaterialTheme.typography.bodyLarge)
+                    Text(
+                        text = temp.label,
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = if (isActive) FontWeight.Bold else FontWeight.Normal,
+                        modifier = Modifier.weight(1f),
+                        color = if (isActive) temp.color else MaterialTheme.colorScheme.onSurface,
+                    )
+                    Text(
+                        text = temp.description,
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.outline,
+                        textAlign = TextAlign.End,
+                    )
+                }
+
+                if (index < ProximityTemperature.entries.toTypedArray().lastIndex) {
+                    HorizontalDivider(
+                        modifier = Modifier.padding(horizontal = 16.dp),
+                        color = MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.5f),
+                    )
+                }
+            }
+        }
+    }
+}
